safari-600.7-branch/LayoutTests/fast/css/counters/after-continuation-expected.txt

This test passes if the numbers displayed below are 1 and 2 in this order.

PASS window.internals.counterValue(document.getElementById('c2')) is '2'

TEST COMPLETE
